+++ modules/config.py
@@ -10,8 +10,13 @@ class Config:
         with open(CONFIG_FILE, "r", encoding="utf-8") as f:
             self.data = json.load(f)
 
-    def get(self, *keys):
+    def get(self, *keys, default=None):
         value = self.data
-        for key in keys:
-            value = value[key]
-        return value
+        try:
+            for key in keys:
+                value = value[key]
+            return value
+        except KeyError:
+            if default is not None:
+                return default
+            raise